requirements.txt
Django==3.1
djangorestframework==3.11.1
numpy
joblib==0.16.0
pandas
